# openapi.gearit.yaml
openapi: 3.1.0
info:
  title: gearit-api
  description: gearit-api
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Аутентификация пользователя
      operationId: login
      security: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type:
                - object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Успешная аутентификация.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
...
  /api/v1/yandex/login:
    get:
      tags:
        - Authentication
      summary: Авторизация через сервер OAuth Yandex.
      operationId: yandex/login
      security: [ ]
      responses:
        200:
          description: Успешное получение редиректа.
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
  /api/v1/yandex/callback:
    get:
      tags:
        - Authentication
      summary: Авторизация через сервер OAuth Yandex.
      operationId: yandex/callback
      security: [ ]
      responses:
        200:
          description: Успешное добавление пользователя в БД и получение токенов доступа.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
...
  /api/v1/automobile:
    post:
      tags:
        - Automobile
      summary: Добавление авто текущего авторизованного пользователя.
      operationId: automobile/add
      security:
        - JWT: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                factoryId:
                  type: integer
                modelId:
                  type: integer
                color:
                  type: string
                license:
                  type: string
                  description: Гос. номер авто.
                bodyType:
                  type: string
                  enum:
                    - SEDAN
                    - HATCHBACK
                    - COUPE
                    - PICKUP
                    - VAN
                    - CROSSOVER
                    - CABRIOLET
                year:
                  type: integer
                  description: Год выпуска.
                odometer:
                  type: integer
                  description: Пробег авто.
      responses:
        200:
          description: Успешное добавление авто для текущего авторизованного пользователя.
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
  /api/v1/automobile/my:
    post:
      tags:
        - Automobile
      summary: Получение списка всех авто текущего авторизованного пользователя.
      operationId: automobile/block:get
      security:
        - JWT: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageableRequest'
      responses:
        200:
          description: Успешное получение списка всех авто текущего авторизованного пользователя.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableResponse'
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
  /api/v1/automobile/delete:
    post:
      tags:
        - Automobile
      summary: Удаление автомобилей текущего авторизованного пользователя.
      operationId: automobile/block:delete
      security:
        - JWT: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                automobileIds:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        200:
          description: Успешное удаление автомобилей текущего авторизованного пользователя.
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
  /api/v1/automobile/factories-models:
    post:
      tags:
        - Automobile
      summary: Получение списка автомобильных производителей и их модели.
      operationId: automobileFactoryModels/get
      security:
        - JWT: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageableRequest'
      responses:
        200:
          description: Успешное получение списка всех автомобильных производителей и их модели.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableResponse'
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
  /api/v1/automobile/{automobileId}:
    post:
      tags:
        - Automobile
      summary: Обновление авто по ID текущего авторизованного пользователя.
      operationId: automobile/update
      security:
        - JWT: [ ]
      parameters:
        - in: path
          name: automobileId
          schema:
            type: integer
          required: true
          description: Идентификатор автомобиля пользователя, который необходимо обновить.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                factoryId:
                  type: integer
                modelId:
                  type: integer
                color:
                  type: string
                license:
                  type: string
                  description: Гос. номер авто.
                bodyType:
                  type: string
                  enum:
                    - SEDAN
                    - HATCHBACK
                    - COUPE
                    - PICKUP
                    - VAN
                    - CROSSOVER
                    - CABRIOLET
                year:
                  type: integer
                  description: Год выпуска.
                odometer:
                  type: integer
                  description: Пробег авто.
      responses:
        200:
          description: Успешное обновление авто для текущего авторизованного пользователя.
        400:
          description: Если были переданы некорректные данные.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
        500:
          description: Ошибка сервера.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientException"
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ClientException:
      type:
        - object
      properties:
        message:
          type: string
          example: 'Incorrect user data'
        extendedHelp:
          type: string
          example: 'Password cannot be null or empty'
    LoginResponse:
      type:
        - object
      properties:
        accessToken:
          type: string
    PageableRequest:
      type:
        - object
      description: Запрос на получение данных с пагинацией, фильтрацией, сортировкой.
      properties:
        page:
          type: integer
          description: Текущая запрашиваемая страница.
        size:
          type: integer
          description: Необходимое количество получаемых элементов.
        container:
          type: array
          description: Контейнер с фильтрами.
          $ref: "#/components/schemas/Filter"
        sort:
          type: object
          properties:
            field:
              type: string
              description: Поле, по которому требуется сортировка.
            direction:
              type: string
              description:
                Направление сортировки.
              enum:
                - ASC
                - DESC
      required:
        - page
        - size
    PageableResponse:
      type:
        - object
      properties:
        page:
          type: integer
          description: Номер текущей страницы.
        pages:
          type: integer
          description: Количество страниц.
        items:
          type: array
        total:
          type: integer
          description:
            Всего элементов.
    Filter:
      type:
        - object
      description:
        Фильтр, используется для формирования выборки по условиям.
      properties:
        field:
          type: string
          description: Наименование поля, по которому предполагается фильтрация
        condition:
          $ref: "#/components/schemas/FilterCondition"
        value:
          type: object
          description: Значение, по которому фильтруются данные.
    FilterCondition:
      type: string
      description:
        Условие фильтрации.
      enum:
        - EQUALS
        - NOT_EQUALS
        - GREATER_THAN
        - LESS_THAN
        - CONTAINS
...